+++ src/args.rs
@@ -2,35 +2,7 @@ use serde::{Deserialize, Serialize};
 
 use display_json::DisplayAsJsonPretty;
 
-use crate::util::{ColorMap1d, ComplexNumber, PostProcessing};
-
-#[derive(Serialize, Deserialize, DisplayAsJsonPretty)]
-pub struct BuddhabrotArgs {
-  #[serde(default = "default_width")]
-  pub width: usize,
-  #[serde(default = "default_height")]
-  pub height: usize,
-  #[serde(default = "default_zoom")]
-  pub zoom: f64,
-  #[serde(default = "default_zpx")]
-  pub zpx: f64,
-  #[serde(default = "default_zpy")]
-  pub zpy: f64,
-  #[serde(default = "default_iter")]
-  pub iter: u32,
-  #[serde(default = "default_buddhabrot_filename")]
-  pub filename: String,
-  #[serde(default)]
-  pub color_map: ColorMap1d,
-  #[serde(default = "default_exponent")]
-  pub exponent: f64,
-  #[serde(default = "default_sample_count")]
-  pub sample_count: u32,
-  #[serde(default)]
-  pub sampler: SamplerArgs,
-  #[serde(default)]
-  pub post_processing: Vec<PostProcessing>,
-}
+use crate::util::{ColorMap1d, ComplexNumber};
 
 #[derive(Serialize, Deserialize, DisplayAsJsonPretty)]
 pub struct JuliaSetArgs {
@@ -66,26 +38,6 @@ pub struct ColorMap1dArgs {
   pub color_map: ColorMap1d,
 }
 
-#[derive(Serialize, Deserialize, DisplayAsJsonPretty)]
-pub struct SamplerArgs {
-  #[serde(default = "default_p_min")]
-  pub p_min: f64,
-  #[serde(default = "default_h")]
-  pub h: f64,
-  #[serde(default = "default_population")]
-  pub population: u32,
-}
-
-impl Default for SamplerArgs {
-  fn default() -> Self {
-    Self {
-      p_min: default_p_min(),
-      h: default_h(),
-      population: default_population(),
-    }
-  }
-}
-
 fn default_width() -> usize {
   1920
 }
@@ -108,30 +60,6 @@ fn default_zpy() -> f64 {
 
 fn default_iter() -> u32 {
   100
-}
-
-fn default_exponent() -> f64 {
-  2.
-}
-
-fn default_sample_count() -> u32 {
-  100_000_000
-}
-
-fn default_p_min() -> f64 {
-  0.5
-}
-
-fn default_h() -> f64 {
-  0.1
-}
-
-fn default_population() -> u32 {
-  1_000_000
-}
-
-fn default_buddhabrot_filename() -> String {
-  "buddhabrot.png".to_owned()
 }
 
 fn default_julia_set_filename() -> String {